Use custom apps for notifications
This adds a macOS and Windows project to handle showing notifications.
It seems the best way to show a notification for each OS is to do so
natively (or close to it in the case of Windows). The macOS project is
native swift done with Xcode and the Windows is in C++ done with Visual
Studio.

I ultimately felt it was easier to control how the notifications worked
by doing it like this. Now Drive backup is using the most modern
APIs to send notifications and can take advantage of this if it is
useful. Since the notifications that are sent from Drive Backup are so
simple, the code in each native app that handles this is also very very
simple.

I'm checking in the Xcode project for macOS into the project's repo
because I am not interested, at this time, in trying to get the bundle
built and signed with GitHub Actions. It seems a lot easier to do
locally with Xcode. In the future I may revisit this, but for now, this
will do. Building the Windows executable is simpler to handle with
GitHub Actions so that I will setup.