Store backup config in file
Each time a backup is run this will store the config in a file. This
file can also be included as a flag to future backups to set the config
to match the previous backup.

This also adds the option to provide a path to where the log file should
be stored.

The flags that dealt with 'logging_' had their names changed to start
with 'log_' to make them shorter.