Move notification trigger into drivebackup.py
Originally I was thinking the notification would be just something done
via the cli (and not a potential gui in the future). But since it seems
to now be working so well, I think it would be appropriate to use
regardless the interface.

The other benefit in doing this is it makes it easy to show a
notification on a failure as well.