Add credentials module and improve handling
This improves the handling of the client credential. Some of the code
for this handling has been moved into it's own module and the parts that
remain in 'drivebackup.py' now have improved error handling.

This also adds the ability to specify the path to the credentials file
in the config (it can be passed in as a command line argument).