Add user commands
This changes the 'cred' command to be the 'user' command and adds
functionality to sign-in, sign-out (was already apart of cred), & view
current user info.

I think it is nice to have the ability to control this via the command
line. I think this will also prove useful if a GUI is implemented.