Add an html success message to authentication
This adds an html success message for when the user successfully
authenticates with Google.

There is a quirk with this, in that you can't use an html message with
the current library, so I modify the necessary function.

Another quirk is with how timeouts work. I would like to use a timeout
on the authentication (300s), but they do not work properly. Rather than
raising a relevant error, an AttributeError is raised because something
else fails when the timeout is reached. So for now, I can't use a
timeout.