Add pin_build.py script
This script replaces the pyproject.toml dependencies with the pinned
dependencies from the lockfile. By default it won't override the
original pyproject.toml, but it can with the '--overwrite' flag.

The idea for this is that it is an easy (the easiest I could come up
with) way to make sure the package will always install the same versions
of **all** of the dependencies. Not just the first order dependencies,
but all of them. This is important since this is an application, not a
library. Compatibility is not a priority (or issue since this will be
installed on its own) but a consistent environment and as reproducible
of a build as possible is. I think this is as close to a 'locked wheel'
distributed via PyPI I am going to get currently.

This will be used in the CI to replace the original pyproject.toml with
the pinned version. It can then build the sdist/bdist for PyPI with it.