Add poetry_plugin_universal2_wheel
This might be overkill...but it works well. In order to build a
universal2 frozen app with PyInstaller, all of the dependencies that are
platform specific must be universal2 (on macOS that is). There were two
issues with this:
1) If a package has x86_64, arm64, & universal2 available, the specific
   wheel for the platform you're running on will get selected, never the
   universal2 wheel.
2) Some packages don't supply universal2 wheels.

So this package automates the process of installing dependencies with
Poetry from the lockfile and making sure any wheels with a shared object
are universal2 wheels on macOS. To summarize how this works:
- Must be installed as a plugin to Poetry
- Will check each dependency from the lockfile and look to see if any of
  them have been installed with a platform specific wheel.
- If a package has a platform specific wheel:
  - First, it will see if there is a universal2 wheel available, and if
    there is it will just download that and replace the platform
    specific wheel with the universal2 version.
  - Second, in lieu of an available universal2 wheel, it will attempt to
    get both types of thin wheels (x86_64 & arm64), fuse them together
    with delocate, & then replace the platform specific wheel with the
    newly fused universal2 version.

Creating this as a Poetry plugin is useful as it gives easier access to
the pyproject.toml and lockfile data. It also automatically runs each
time dependencies are installed.