Add poetry-plugin-lock-build
Instead of using a python script to manually replace the
'pyproject.toml' dependencies, this adds a Poetry plugin to do the work.

A major drawback of the other approach was how it would make a new file
that wasn't actually the 'pyproject.toml' or would overwrite the
'pyproject.toml'. Overwriting would be okay in the CI, but for testing
locally this would not be acceptable. This meant I would have had to
manually rename the pinned toml with the original when testing builds.
Not the end of the world, but not ideal.

This super duper plugin runs before building if you include an empty
table for it in the 'pyproject.toml'. It adds all the necessary
dependencies to support the main dependencies. It uses the versions from
the lockfile to pin each dependency in the build output. It also makes
sure to include any necessary markers.

Most of the heavy lifting for this is still being handled by
'poetry-plugin-export'. Because of this it really should work almost the
same (a little better actually) as the previous approach of parsing the
text output from 'poetry-plugin-export' and manually writing it back
into 'pyproject.toml'.