Rename poetry-plugin-lock-build
This renames poetry-plugin-lock-build to poetry-plugin-pin-build. After
learning a little more about python packaging reproducibility, I think
it is more accurate to not call this "lock-build". This only pins the
dependencies, but does not verify anything further than that (i.e.
hashes). This is still perfectly fine for my use case, but just isn't
accurate to label it "lock".