Add hooks for PyInstaller
These hooks configure what PyInstaller installs in the frozen build.
For our own package, we need the metadata so we can display our version.
We also need to tell it to include the resource directory files. We
could do this via the command line but since we need the hook anyway, we
might as well include that.

There is also a hook for googleapiclient.model which is used to thin out
the 'discovery_cache' directory of data files. This directory is pretty
big (>70MB) and it just includes configuration for different Google
apis. We only need the 'drive.v3.json' configuration, so by omitting the
rest we greatly reduce the size of the resulting build.