Start adding build GitHub Action
This gets a first pass at building the macOS PyInstaller binary. I kept
debating if it would be better to write 3 separate jobs, or 1 matrix job
for building the binaries. I'm split because it seems there are just
pros/cons both ways. But I'm actually surprised how little path code I
was able to get in the action (which is good). I would've been okay
using Powershell on Windows but it would be too tricky to run the same
shell code on bash/Powershell...which is probably obvious. I should give
the Windows bash a shot. The main reason I don't want to write 3
separate jobs is because it will have a lot of overlapping code, I'd
rather not have to write much of the job multiple times.