Fix notification buidl path outputs
First problem: Using an absolute path was preventing the build path from
being portable because the workspace is in a different place on each
difference OS. Just use a relative path.

Second problem: I was using a windows path in the linux runner trying to
build the package. This was clearly not going to work. Solution is to
also output the posix build path from the windows-notifications job.