Simplify app download, use temp dir & pipx default
This simplifies the artifact download for the package build. Since the
only two artifacts that exist we do want to download, it is just easier
to let download-artifact download 'everything'. Plus, if this were to
change, we now download into a temporary directory, so there shouldn't
be any issues.

This also sets the default pipx python executable for the binary build
jobs.