Remove absolute path for notif matrix build path
When setting the notification build paths in the matrix, the workspace
path is different when it is set compared to after running
'actions/checkout'. Because of this we are installing into the wrong
place. The simplest fix is to just use relative paths. While I tried my
hardest to avoid this in the notifications workflow, I think it
shouldn't be an issue here.