Fix macos codesign and archive
This fixes a code signing error from PyInstaller. Codesign was creating
a _CodeSignature Directory for the Python executable, but PyInstaller
was not copying it into the build. We manually codesign that executable,
which again creates the _CodeSignature directory, to fix this.

This also fixes how zips were being created on mac (which is needed for
notarization). The zip was not maintaining symlinks, it was replacing
them with the linked file. This was also causing codesigning problems
and I'm assuming would have caused other issues at some point. Now we
use ditto directly to create zip archives on macos.