Fix pyinstaller build command
While the previous build command worked fine on mac & linux, it was
not working on windows. It turns out the problem was with how shlex was
parsing the string in 'split'. It was removing the '\', which completely
messed up the distpath. This commit fixes that by putting distpath in
quotes. shlex.split will not leave the backslashes alone.

This also improves the script argument by making sure it too will have
platform appropriate separators.