Add notification directory when adding for windows
When copying the windows notification app into the app binary directory
we have to first make the necessary parent directories. This problem
doesn't exist on the mac version because for that we need to use
'copytree' which creates the directories for us. Since for the windows
version we are only copying one file we use 'copy2', which does not make
the directories for us.