Switch to unzip to unarchive notification zip
shutil.unpack_archive does not preserve file permissions when
unpacking zip archives. This causes problems with the mac notifications
app. To my happy surprise, the 'unzip' tool is available on both the
macos and windows runner. On Windows it is there because we are using
git bash...so +1 for git bash.