Add release workflow and publish to test pypi
This adds a workflow that will run on releases. It uploads release
assets and publishes the python package to pypi (wheel and sdist).

This also adds a job to the build workflow to publish the python
package on test pypi. This is something that is described in the python
packaging user guide:
https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/