Make publishing to Test PyPI its own workflow
If the same version is published to the package index it fails because
you can't overwrite or replace what is already there. This would be an
issue if I was trying to publish to Test PyPI with every commit. Because
of this, that job needs to be removed from the workflow. However, I like
the idea of being able to publish to Test PyPI, for testing, so that job
is now in its own workflow that can be triggered manually.