Check whether to run poetry plugins in dispatch
Rather than checking if we should potentially run the poetry plugins in
the 'activate' function, check in the dispatched function. When
'poetry.pyproject' is queried, if the pyproject.toml doesn't exist, it
exits, and does nothing. This will happen for *all* functions. Even for
'init' and 'new'. This is obviously problematic. By checking in the
dispatched function, *after* we check if the corresponding command was
run (i.e. build or install), we avoid this problem. We won't ever ask
when functions are run where not having a pyproject.toml is okay. We
only ask for build and install, where you already need the
pyproject.toml.